+++ app/src/main/java/jp/ac/uryukyu/ie/e205738/LivingThing.java
@@ -26,7 +26,7 @@ public class LivingThing {
             opponent.wounded(damage);
         }
     }
-    public void wounded(int damage){
+    public void wounded(double damage){
         this.hitPoint -= damage;
         if( this.hitPoint < 0 ) {
             dead = true;
